requirements.txt
openstacksdk>=3.0.0
rich>=13.0.0
